Rename Bxstm namespace to NintendoWare
