+++ src/VGAudio.Uwp/Audio/ContainerTypes.cs
@@ -4,12 +4,12 @@ using System.IO;
 using System.Linq;
 using VGAudio.Containers;
 using VGAudio.Containers.Adx;
-using VGAudio.Containers.Bxstm;
 using VGAudio.Containers.Dsp;
 using VGAudio.Containers.Genh;
 using VGAudio.Containers.Hca;
 using VGAudio.Containers.Hps;
 using VGAudio.Containers.Idsp;
+using VGAudio.Containers.NintendoWare;
 using VGAudio.Containers.Wave;
 
 namespace VGAudio.Uwp.Audio
