+++ src/VGAudio/Containers/NintendoWare/BcstmReader.cs
@@ -1,8 +1,8 @@
 ﻿using System.IO;
-using VGAudio.Containers.Bxstm.Structures;
+using VGAudio.Containers.NintendoWare.Structures;
 using VGAudio.Formats;
 
-namespace VGAudio.Containers.Bxstm
+namespace VGAudio.Containers.NintendoWare
 {
     public class BcstmReader : AudioReader<BcstmReader, BxstmStructure, BxstmConfiguration>
     {
