+++ src/VGAudio/Containers/NintendoWare/BfstmWriter.cs
@@ -1,7 +1,7 @@
 ﻿using System.IO;
 using VGAudio.Formats;
 
-namespace VGAudio.Containers.Bxstm
+namespace VGAudio.Containers.NintendoWare
 {
     public class BfstmWriter : AudioWriter<BfstmWriter, BxstmConfiguration>
     {
